Cleanup of code based on CLion analysis

+++ UbiForm/ResourceDiscovery/ResourceDiscoveryStore.cpp
@@ -83,6 +83,6 @@ SocketMessage *ResourceDiscoveryStore::generateRDResponse(SocketMessage *sm, Res
 }
 
 ResourceDiscoveryStore::ResourceDiscoveryStore(SystemSchemas & ss) : systemSchemas(ss), generator() {
-    unsigned randomSeed = std::chrono::system_clock::now().time_since_epoch().count();
+    long randomSeed = std::chrono::system_clock::now().time_since_epoch().count();
     generator.seed(randomSeed);
 }
